Move sbom upload to sub commands directory
